# lib/api-spec/openapi.yaml
openapi: 3.1.0
info:
  # Do not change the title, if the title changes, the import paths will be broken
  title: Api
  version: 0.1.0
  description: Neel Kamal Homestay API
servers:
  - url: /api
    description: Base API path
...
paths:
...
  /rooms:
    get:
      operationId: listRooms
      tags: [rooms]
      summary: List rooms
      parameters:
        - in: query
          name: admin
          schema:
            type: string
          description: If "true", return all rooms including hidden
      responses:
        "200":
          description: List of rooms
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/Room"

    post:
      operationId: createRoom
      tags: [rooms]
      summary: Create a room (admin)
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "201":
          description: Room created
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/{id}:
    get:
      operationId: getRoom
      tags: [rooms]
      summary: Get a single room
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Room details
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
        "404":
          description: Not found
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/ErrorResponse"

    put:
      operationId: updateRoom
      tags: [rooms]
      summary: Update a room (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomInput"
      responses:
        "200":
          description: Room updated
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"

    delete:
      operationId: deleteRoom
      tags: [rooms]
      summary: Delete a room (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Deleted
  /rooms/{id}/toggle-available:
    patch:
      operationId: toggleRoomAvailable
      tags: [rooms]
      summary: Toggle room availability (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/{id}/toggle-visible:
    patch:
      operationId: toggleRoomVisible
      tags: [rooms]
      summary: Toggle room visibility (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Toggled
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/Room"
  /rooms/{id}/images:
    get:
      operationId: listRoomImages
      tags: [rooms]
      summary: List images for a room
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Room images
          content:
            application/json:
              schema:
                type: array
                items:
                  $ref: "#/components/schemas/RoomImage"

    post:
      operationId: addRoomImage
      tags: [rooms]
      summary: Associate a media item with a room (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
      requestBody:
        required: true
        content:
          application/json:
            schema:
              $ref: "#/components/schemas/RoomImageInput"
      responses:
        "201":
          description: Image added
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomImage"
  /rooms/{id}/images/{imageId}:
    delete:
      operationId: removeRoomImage
      tags: [rooms]
      summary: Remove an image from a room (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: imageId
          required: true
          schema:
            type: integer
      responses:
        "204":
          description: Removed
  /rooms/{id}/images/{imageId}/set-cover:
    patch:
      operationId: setRoomCoverImage
      tags: [rooms]
      summary: Set cover image for a room (admin)
      parameters:
        - in: path
          name: id
          required: true
          schema:
            type: integer
        - in: path
          name: imageId
          required: true
          schema:
            type: integer
      responses:
        "200":
          description: Cover set
          content:
            application/json:
              schema:
                $ref: "#/components/schemas/RoomImage"

  # ─── MEDIA ───────────────────────────────────────────────────────────────────
...
components:
  schemas:
...
    ErrorResponse:
      type: object
      properties:
        error:
          type: string
      required: [error]
...
    Room:
      type: object
      properties:
        id:
          type: integer
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        pricePerNight:
          type: ["number", "null"]
        maxGuests:
          type: integer
        bedType:
          type: string
        amenities:
          type: array
          items:
            type: string
        isAvailable:
          type: boolean
        isVisible:
          type: boolean
        coverImageUrl:
          type: ["string", "null"]
        sortOrder:
          type: integer
        createdAt:
          type: string
        updatedAt:
          type: string
      required:
        - id
        - name
        - slug
        - description
        - shortDescription
        - maxGuests
        - bedType
        - amenities
        - isAvailable
        - isVisible
        - sortOrder
        - createdAt
        - updatedAt
    RoomInput:
      type: object
      properties:
        name:
          type: string
        slug:
          type: string
        description:
          type: string
        shortDescription:
          type: string
        pricePerNight:
          type: ["number", "null"]
        maxGuests:
          type: integer
        bedType:
          type: string
        amenities:
          type: array
          items:
            type: string
        isAvailable:
          type: boolean
        isVisible:
          type: boolean
        sortOrder:
          type: integer
      required: [name, slug, description, shortDescription, maxGuests, bedType]
    RoomImage:
      type: object
      properties:
        id:
          type: integer
        roomId:
          type: integer
        mediaId:
          type: integer
        isCover:
          type: boolean
        sortOrder:
          type: integer
        url:
          type: string
        altText:
          type: ["string", "null"]
        filename:
          type: string
      required: [id, roomId, mediaId, isCover, sortOrder, url, filename]
    RoomImageInput:
      type: object
      properties:
        mediaId:
          type: integer
        isCover:
          type: boolean
        sortOrder:
          type: integer
      required: [mediaId]

    # ── Media ─────────────────────────────────────────────────────────────────
...