# lib/api-zod/src/generated/types/room.ts
/**
 * Generated by orval v8.21.0 🍺
 * Do not edit manually.
 * Api
 * Neel Kamal Homestay API
 * OpenAPI spec version: 0.1.0
 */

export interface Room {
  id: number;
  name: string;
  slug: string;
  description: string;
  shortDescription: string;
  /** @nullable */
  pricePerNight?: number | null;
  maxGuests: number;
  bedType: string;
  amenities: string[];
  isAvailable: boolean;
  isVisible: boolean;
  /** @nullable */
  coverImageUrl?: string | null;
  sortOrder: number;
  createdAt: string;
  updatedAt: string;
}